BMD-6423: lint
